// packages/labextension/src/model/assignmentDetail.ts
/**
 * Grader Extension API Schemas
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AssignmentSettings } from './assignmentSettings';
import { Submission } from './submission';

export interface AssignmentDetail {
  id?: number;
  name?: string;
  settings?: AssignmentSettings;
  status?: AssignmentDetail.StatusEnum;
  points?: number;
  submissions?: Array<Submission>;
}
export namespace AssignmentDetail {
  export const StatusEnum = {
    Created: 'created',
    Pushed: 'pushed',
    Released: 'released',
    Complete: 'complete'
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
}